:recycle: Update detectionOn.sh for better code quality
- Replace `-o` with `||` for better readability and consistency
- Add shellcheck directives for better code quality
- Improve variable handling with `${var:?}` to prevent errors
- Update conditional statements for better readability

+++ firmware_mod/scripts/detectionOn.sh
@@ -17,8 +17,9 @@ debug_msg () {
 send_snapshot() {
 
 	# Publish a mqtt message
-	if [ "$publish_mqtt_message" = true -o "$publish_mqtt_snapshot" = true ] ; then
+	if [ "$publish_mqtt_message" = true ] || [ "$publish_mqtt_snapshot" = true ] ; then
 		(
+                # shellcheck source=config/mqtt.conf.dist
 		. /system/sdcard/config/mqtt.conf
 
 		if [ "$publish_mqtt_message" = true ] ; then
@@ -41,12 +42,13 @@ send_snapshot() {
 	# Send a telegram message
 	if [ "$send_telegram" = true ]; then
 		(
+		# shellcheck source=config/telegram.conf.dist
 		include /system/sdcard/config/telegram.conf
 
 		if [ "$telegram_alert_type" = "text" ] ; then
 			debug_msg "Send telegram text"
 			/system/sdcard/bin/telegram m "Motion detected"
-		elif [ "$telegram_alert_type" = "image" -o "$telegram_alert_type" = "video+image" ] ; then
+		elif [ "$telegram_alert_type" = "image" ] || [ "$telegram_alert_type" = "video+image" ] ; then
 			debug_msg "Send telegram image"
 			/system/sdcard/bin/telegram p "$snapshot_tempfile"
 		fi
@@ -61,7 +63,7 @@ send_snapshot() {
 		if [ "$matrix_alert_type" = "text" ] ; then
 			debug_msg "Send matrix text"
 			/system/sdcard/bin/matrix m "Motion detected"
-		elif [ "$matrix_alert_type" = "image" -o "$matrix_alert_type" = "video+image" ] ; then
+		elif [ "$matrix_alert_type" = "image" ] || [ "$matrix_alert_type" = "video+image" ] ; then
 			debug_msg "Send matrix image"
 			/system/sdcard/bin/matrix p "$filename" "$snapshot_tempfile"
 		fi
@@ -113,7 +115,7 @@ send_snapshot() {
 
 		# Limit the number of snapshots
 		if [ "$(ls "$save_snapshot_dir" | wc -l)" -ge "$max_snapshot_days" ]; then
-			rm -rf "${save_snapshot_dir}/$(ls -ltr "$save_snapshot_dir" | awk 'NR==2{print $9}')"
+			rm -rf "${save_snapshot_dir:?}/$(ls -ltr "$save_snapshot_dir" | awk 'NR==2{print $9}')"
 		fi
 
 		chmod "$save_files_attr" "$snapshot_tempfile"
@@ -225,8 +227,8 @@ send_snapshot &
 if [ "$save_video" = true ] ||
    [ "$smb_video" = true ] ||
    [ "$dropbox_video" = true ] ||
-  ([ "$send_telegram" = true ] && ([ "$telegram_alert_type" = video+image ] || [ "$telegram_alert_type" = video ])) ||
-  ([ "$send_matrix" = true ] && ([ "$matrix_alert_type" = video+image ] || [ "$matrix_alert_type" = video ])) ||
+ { [ "$send_telegram" = true ] && { [ "$telegram_alert_type" = video+image ] || [ "$telegram_alert_type" = video ]; }; } ||
+ { [ "$send_matrix" = true ] && { [ "$matrix_alert_type" = video+image ] || [ "$matrix_alert_type" = video ]; }; } ||
    [ "$publish_mqtt_video" = true ]
 then
 	record_video
@@ -246,7 +248,7 @@ if [ "$save_video" = true ] ; then
 
 	# Limit the number of videos
 	if [ "$(ls "$save_video_dir" | wc -l)" -ge "$max_video_days" ]; then
-		rm -rf "$save_video_dir/$(ls -ltr "$save_video_dir" | awk 'NR==2{print $9}')"
+		rm -rf "${save_video_dir:?}/$(ls -ltr "$save_video_dir" | awk 'NR==2{print $9}')"
 	fi
 
 	chmod "$save_files_attr" "$video_tempfile"
@@ -356,7 +358,7 @@ if [ "$send_telegram" = true ]; then
 	(
 	include /system/sdcard/config/telegram.conf
 
-	if [ "$telegram_alert_type" = "video" -o "$telegram_alert_type" = "video+image" ] ; then
+	if [ "$telegram_alert_type" = "video" ] || [ "$telegram_alert_type" = "video+image" ] ; then
 		if [ "$video_use_rtsp" = true ]; then
 			if [ "$AUDIOFORMAT" = "PCMU" ] || [ "$AUDIOFORMAT" = "OFF" ] ; then
 				# Convert file to mp4 and remove audio stream so video plays in telegram app
@@ -383,7 +385,7 @@ if [ "$send_matrix" = true ]; then
 	(
 	include /system/sdcard/config/matrix.conf
 
-	if [ "$matrix_alert_type" = "video" -o "$matrix_alert_type" = "video+image" ] ; then
+	if [ "$matrix_alert_type" = "video" ] || [ "$matrix_alert_type" = "video+image" ] ; then
 		debug_msg "Send matrix video"
 		if [ "$video_use_rtsp" = true ]; then
 			/system/sdcard/bin/matrix v "$filename" "$video_tempfile"
